Two Samples Excel Projects From Work
Two excel projects (one with macros enabled). Produced and reproduced
here with permission from my supervisor.  Values/names have been changed
for the BACKEND project as it is a sensitive internal project,
UNIVERSITY project has not changed values because all information for
that project was discovered through the public domain.